pyproject.toml
[tool.poetry]
name = "eu-ai-act-chatbot"
version = "0.1.0"
description = "EU AI Act compliance chatbot using hybrid retrieval"
authors = ["Agentic Architects <agentic_architects@thoughtgeek.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
unstructured = {extras = ["pdf"], version = "^0.14.10"} # Added pdf extra for partition_pdf
pinecone-client = "^3.2.2" # Updated to v3 syntax
sentence-transformers = "^2.7.0"
neo4j = "^5.22.0"
requests = "^2.32.3"
langchain = "^0.2.10" # Updated langchain
langchain-community = "^0.2.7" # Updated langchain-community
python-dotenv = "^1.0.1"
pydantic = "^2.8.2"
fastapi = "^0.111.1"
uvicorn = {extras = ["standard"], version = "^0.30.1"} # Added standard extra
python-multipart = "^0.0.9"
mangum = "^0.17.0"  # For AWS Lambda integration
openai = "^1.30.0" # Use OpenAI SDK instead of openrouter library
langchain-openai = "^0.1.17" # Needed for some LangChain integrations, good to have

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.1"
black = "^24.4.2"
isort = "^5.13.2"
pylint = "^3.2.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 